--- a/notifications_utils/version.py
+++ b/notifications_utils/version.py
@@ -5,4 +5,4 @@
 # - `make version-minor` for new features
 # - `make version-patch` for bug fixes
 
-__version__ = '51.1.1'  # 5fd0bbc81c3653cee9fa2ada1b78c8c5
+__version__ = '51.2.0'  # 1e1f0a68e40b7c2b164567e4d7e1d373