https://www.pivotaltracker.com/story/show/177535141

This PR does two things:

- Introduces timeout functionality to avoid RecipientCSV
getting stuck processing massive CSVs.

- Rewrites / removes some the tests for this class, partly
to avoid the timeout but also just make it faster!

Please review commit by commit.

## Test timing stats

```
# before
time pytest
pytest  49.66s user 2.62s system 97% cpu 53.573 total

# after
time pytest
pytest  28.59s user 2.46s system 96% cpu 32.184 total
```

## Request timings

These are for [previewing an uploaded CSV in the Admin app](https://github.com/alphagov/notifications-admin/blob/925f86aa705583e920ea4f0300c3f85fd8546c90/app/main/views/send.py#L645).

Note: these might be a bit dodgy because there's no way to filter for the "/check/" URLs in Kibana: you can't search / filter for non-alphanumeric characters in strings. The query in the screenshot seems to filter out everything we don't care about. I might try and get the same data out of Athena, but I doubt it will change the default.

<img width="1117" alt="Screenshot 2021-12-07 at 16 41 36" src="https://user-images.githubusercontent.com/9029009/145070012-36f65b4a-9939-47e3-8067-8ff228930e4c.png">