@benthorner 2021-12-08
Data from Athena for the past month shows the same trend, with only a handful of requests taking over 10s:

<img width="647" alt="Screenshot 2021-12-08 at 10 24 46" src="https://user-images.githubusercontent.com/9029009/145192224-358812f2-4336-4cd9-bfef-969942672bde.png">

(X-axis: request time in seconds, Y-axis: count)